+++ enable-bbp-block-theme.php
@@ -3,7 +3,7 @@
  * Plugin Name:  Enable bbPress for Block Themes
  * Plugin URI:   https://www.webmandesign.eu/portfolio/enable-bbp-block-theme-wordpress-plugin/
  * Description:  Enables bbPress for a full site editing WordPress block theme.
- * Version:      1.0.1
+ * Version:      1.1.0
  * Author:       WebMan Design, Oliver Juhas
  * Author URI:   https://www.webmandesign.eu/
  * License:      GPL-3.0-or-later
@@ -11,8 +11,8 @@
  * Text Domain:  enable-bbp-block-theme
  * Domain Path:  /languages
  *
- * Requires PHP:       7.0
- * Requires at least:  6.7
+ * Requires PHP:       8.0
+ * Requires at least:  7.0
  *
  * GitHub Plugin URI:  https://github.com/webmandesign/enable-bbp-block-theme
  *
@@ -25,14 +25,19 @@
  * @package  Enable bbPress for Block Themes
  */
 
+namespace WebManDesign\bbPress\Block_Theme;
+
 // Exit if accessed directly.
 defined( 'ABSPATH' ) || exit;
 
 // Constants.
-define( 'WMD_EBBT_NAMESPACE', 'WebManDesign\bbPress\Block_Theme' );
-define( 'WMD_EBBT_FILE', __FILE__ );
-define( 'WMD_EBBT_PATH', plugin_dir_path( WMD_EBBT_FILE ) ); // Trailing slashed.
+
+	const WMD_EBBT_NAMESPACE = __NAMESPACE__;
+	const WMD_EBBT_FILE      = __FILE__;
+	const WMD_EBBT_PATH      = trailingslashit( plugin_dir_path( WMD_EBBT_FILE ) );
 
 // Load the functionality.
-require_once WMD_EBBT_PATH . 'includes/Autoload.php';
-WebManDesign\bbPress\Block_Theme\Load::init();
+
+	require_once WMD_EBBT_PATH . 'includes/Autoload.php';
+
+	Load::init();