+++ changelog.md
@@ -1,5 +1,17 @@
 # Enable bbPress for Block Themes changelog
 
+## 1.1.0, 20260918
+
+### Fixed
+- Compatibility with bbPress 2.6.17 (Making sure the original bbPress template hierarchy is preserved.)
+
+### File updates
+	changelog.md
+	enable-bbp-block-theme.php
+	readme.txt
+	includes/Compatibility.php
+
+
 ## 1.0.1, 20250804
 
 ### Fixed